+++ packages/import-unlayer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @templatical/import-unlayer
 
+## 0.13.0
+
+### Patch Changes
+
+- @templatical/types@0.13.0
+
 ## 0.12.1
 
 ### Patch Changes